+++ time_authority/__init__.py
@@ -1,0 +1,2 @@
+# Time Authority package
+__all__ = []